Create week1
Welcome to your first lab on fixing problems in Python. In this lab, you'll first have to fix an incorrect Python script. This includes:

Fixing the file permissions to make it executable.
Fixing a bug in the code.
After that, you'll write your own Python module and use it from the original script.